+++ src/main/java/blackjack/domain/Participant.java
@@ -2,6 +2,9 @@ package blackjack.domain;
 
 import java.util.List;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Cards;
+
 public class Participant {
 	private static final String DEALER_NAME = "딜러";
 	private static final String BLANK = "";
@@ -10,43 +13,43 @@ public class Participant {
 	private static final String ERROR_MESSAGE_UNAVAILABLE_CHARACTER = "[ERROR] 이름에 특수문자가 포함될 수 없습니다.";
 	private static final int MAX_SCORE = 21;
 
-    private final String name;
-    private final Cards cards;
+	private final String name;
+	private final Cards cards;
 
-    private Participant(String name) {
-        this.name = name;
-        this.cards = new Cards();
-    }
+	private Participant(String name) {
+		this.name = name;
+		this.cards = new Cards();
+	}
 
-    public static Participant createDealer() {
-        return new Participant(DEALER_NAME);
-    }
+	public static Participant createDealer() {
+		return new Participant(DEALER_NAME);
+	}
 
-    public static Participant createPlayer(String input) {
-        return new Participant(validateName(input.trim()));
-    }
+	public static Participant createPlayer(String input) {
+		return new Participant(validateName(input.trim()));
+	}
 
-    private static String validateName(String input) {
-        checkBlankName(input);
-        checkUnavailableName(input);
-        return input;
-    }
+	private static String validateName(String input) {
+		checkBlankName(input);
+		checkUnavailableName(input);
+		return input;
+	}
 
-    private static void checkBlankName(String input) {
-        if (input.equals("")) {
-            throw new IllegalArgumentException(ERROR_MESSAGE_EMPTY_NAME);
-        }
-    }
+	private static void checkBlankName(String input) {
+		if (input.equals(BLANK)) {
+			throw new IllegalArgumentException(ERROR_MESSAGE_EMPTY_NAME);
+		}
+	}
 
-    private static void checkUnavailableName(String input) {
-        if (input.matches(ALLOWED_CHARACTERS)) {
-            throw new IllegalArgumentException(ERROR_MESSAGE_UNAVAILABLE_CHARACTER);
-        }
-    }
+	private static void checkUnavailableName(String input) {
+		if (input.matches(ALLOWED_CHARACTERS)) {
+			throw new IllegalArgumentException(ERROR_MESSAGE_UNAVAILABLE_CHARACTER);
+		}
+	}
 
-    public void receiveCard(Card card) {
-        cards.addCard(card);
-    }
+	public void receiveCard(Card card) {
+		cards.addCard(card);
+	}
 
 	public Boolean isOverMaxScore() {
 		return getScore() >= MAX_SCORE;
@@ -56,11 +59,11 @@ public class Participant {
 		return cards.sum();
 	}
 
-    public String getName() {
-        return name;
-    }
+	public String getName() {
+		return name;
+	}
 
-    public List<Card> getCards() {
-        return cards.getCards();
-    }
+	public List<Card> getCards() {
+		return cards.getCards();
+	}
 }
