+++ src/main/java/blackjack/domain/card/Card.java
@@ -1,6 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public class Card {
+	private static final int ACE = 1;
+
 	private final String name;
 	private final int value;
 
@@ -9,15 +11,15 @@ public class Card {
 		this.value = value;
 	}
 
+	public boolean isAce() {
+		return value == ACE;
+	}
+
 	public String getName() {
 		return name;
 	}
 
 	public int getValue() {
 		return value;
-	}
-
-	public boolean isAce() {
-		return value == 1;
 	}
 }
