+++ src/main/java/blackjack/view/ResultView.java
@@ -1,23 +1,26 @@
 package blackjack.view;
 
-import blackjack.domain.dto.BlackJackDto;
 import blackjack.domain.Participant;
+import blackjack.domain.dto.BlackJackDto;
 
 public class ResultView {
 
-	private static final String RESULT_DELIMITER = " - 결과: ";
+	private static final String MESSAGE_HAND_OUT_CARD = "%n딜러와 %s 에게 2장의 카드를 나누었습니다.%n";
+	private static final String NAME_DELIMITER = ", ";
+	private static final String MESSAGE_SCORE_OVER_21 = "가진 카드의 합이 21 이상이 되어 더 이상 카드를 받을 수 없습니다.";
 	private static final String MESSAGE_DEALER_NOT_RECEIVE = "딜러는 16이하라 한장의 카드를 더 받았습니다";
 	private static final String MESSAGE_DEALER_RECEIVE = "딜러는 17이상이라 카드를 더 받지 않았습니다.";
-	private static final String NAME_DELIMITER = ", ";
-	private static final String MESSAGE_HAND_OUT_CARD = "%n딜러와 %s 에게 2장의 카드를 나누었습니다.%n";
-	private static final String MESSAGE_SCORE_OVER_21 = "가진 카드의 합이 21 이상이 되어 더 이상 카드를 받을 수 없습니다.";
+	private static final String RESULT_DELIMITER = " - 결과: ";
+	private static final String MESSAGE_FINAL_RESULT = "\n## 최종 승패";
 
 	public static void showStartingStatus(BlackJackDto blackJackDto) {
 		String[] playerNames = blackJackDto.getPlayers().stream()
 			.map(Participant::getName)
 			.toArray(String[]::new);
+
 		System.out.printf(MESSAGE_HAND_OUT_CARD, String.join(NAME_DELIMITER, playerNames));
 		System.out.println(blackJackDto.getDealerOpenCard());
+
 		for (Participant player : blackJackDto.getPlayers()) {
 			showEachPlayerCurrentStatus(blackJackDto, player);
 		}
@@ -54,7 +57,7 @@ public class ResultView {
 	}
 
 	public static void showResult(BlackJackDto blackJackDto) {
-		System.out.println("\n## 최종 승패");
+		System.out.println(MESSAGE_FINAL_RESULT);
 		System.out.println(blackJackDto.getDealerResult());
 		blackJackDto.getPlayersResult().forEach(System.out::println);
 	}
