+++ src/test/java/blackjack/domain/ParticipantTest.java
@@ -1,8 +1,7 @@
-package blackjack;
+package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
-import blackjack.domain.Participant;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
