+++ src/test/java/blackjack/domain/card/CardDeckTest.java
@@ -1,9 +1,9 @@
-package blackjack;
+package blackjack.domain.card;
 
 import static org.assertj.core.api.Assertions.*;
 
-import blackjack.domain.Card;
-import blackjack.domain.CardDeck;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardDeck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
